requirements.txt
aiogram
asyncio
python-telegram-bot==20.4
